Characters + religious diversity
I've added Madagascar's characters and diversified the state + pop religions. I have also made a bigger change in the addition of the "autarkic" ideology, an ideology that supports isolationism and closed borders that should appear among unrecognized states, traditionalist states, and also among happy command economists, socialists, and angry populists, + extra weight for East Asia if China gets wrecked in the opium war. Ranavalona starts with this ideology to start. The weights for the ideology are not balanced or tested at all right now.