Cali natives and Madasgascar fixes
I accidentally deleted the definitions of all my tags in Madagascar lol. Also fixed the spline stuff there, i did it wrong before. I also un-peninsula'd vancouver and modified the heightmap in Madagascar a bit for the sake of a railroad - you can definitely see the "scar" of me doing it manually in both places