Canadian territories redraw, HBC boundaries, BC, Alaska, Yukon, Montana natives
This took a lot longer than the others. The HBC is really difficult to draw as a country seeing as it was basically a bunch of supply depots and rivers. I also removed all the parentheses in the states file, though that was incidental. Looking forward to getting back into more consistent progress now that the HBC is like 80% done. The three northern territories were drawn weirdly, seemingly in service of giving them port cities that they never actually got, so I've redrawn them to be more accurate, as well - if in the end we decide Yukon does need a port (perhaps for fish), we'll stick it on a lake or something; they really don't need a shipping port.