attempts at bugfixing
Some of the issues with current script may not be solvable ("weight=" may not actually meaningfully work in random effects, for instance, and the global_list stuff I may have utterly misunderstood). Other issues are completely inexplicable and maddening. I do not understand why the event I specifically say to trigger for the USA just does not. This'll take a while.